[Whisper] Fix float32/float16 dtype mismatch in integration tests
whisper-large-v3 and distil-large-v3 are now saved as float16 on the Hub.
Tests loaded the model without explicit dtype (loaded as float16) but cast
input_features only to device (stayed float32), causing:
  RuntimeError: Input type (float) and bias type (c10::Half) should be the same

Fix: use input_features.to(torch_device, dtype=model.dtype) in 5 tests:
- test_large_generation
- test_large_generation_multilingual
- test_large_batched_generation
- test_distil_token_timestamp_generation
- test_large_timestamp_generation

+++ tests/models/whisper/test_modeling_whisper.py
@@ -1406,7 +1406,7 @@ class WhisperModelIntegrationTests(unittest.TestCase):
 
         input_speech = self._load_datasamples(1)
         input_features = processor(input_speech, return_tensors="pt", sampling_rate=16_000).input_features
-        input_features = input_features.to(torch_device)
+        input_features = input_features.to(torch_device, dtype=model.dtype)
 
         generated_ids = model.generate(
             input_features, do_sample=False, max_length=20, language="<|en|>", task="transcribe"
@@ -1427,7 +1427,7 @@ class WhisperModelIntegrationTests(unittest.TestCase):
 
         input_speech = next(iter(ds))["audio"]["array"]
         input_features = processor(input_speech, return_tensors="pt", sampling_rate=16_000).input_features
-        input_features = input_features.to(torch_device)
+        input_features = input_features.to(torch_device, dtype=model.dtype)
 
         generated_ids = model.generate(
             input_features, do_sample=False, max_length=20, language="<|de|>", task="transcribe"
@@ -1452,7 +1452,7 @@ class WhisperModelIntegrationTests(unittest.TestCase):
 
         input_speech = self._load_datasamples(4)
         input_features = processor(input_speech, return_tensors="pt", sampling_rate=16_000).input_features
-        input_features = input_features.to(torch_device)
+        input_features = input_features.to(torch_device, dtype=model.dtype)
         generated_ids = model.generate(input_features, max_length=20, task="translate")
 
         # fmt: off
@@ -1632,7 +1632,7 @@ class WhisperModelIntegrationTests(unittest.TestCase):
 
         input_speech = np.concatenate(self._load_datasamples(4))
         input_features = processor(input_speech, return_tensors="pt", sampling_rate=16_000).input_features
-        input_features = input_features.to(torch_device)
+        input_features = input_features.to(torch_device, dtype=model.dtype)
 
         _ = model.generate(
             input_features, max_length=448, return_timestamps=True, return_token_timestamps=True, return_segments=True
@@ -1903,7 +1903,7 @@ class WhisperModelIntegrationTests(unittest.TestCase):
             return_tensors="pt",
             sampling_rate=16_000,
         ).input_features
-        input_features = input_features.to(torch_device)
+        input_features = input_features.to(torch_device, dtype=model.dtype)
 
         generated_ids = model.generate(
             input_features, max_length=448, return_timestamps=True, condition_on_prev_tokens=True
